# api/sarcoma.openapi.yaml
openapi: 3.0.3
info:
  title: Sarcom FastTrack API
  version: 1.0.0
  description: Go WebAPI for Sarcom FastTrack patient handoff reports.
  contact:
    email: support@sarcoma-fasttrack.local
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - description: Local API
    url: http://localhost:8000
  - description: WAC Gateway API route
    url: /sarcoma-fasttrack-api
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/reports:
    get:
      tags: [reports]
      summary: List reports visible to the current user.
      operationId: listReports
      responses:
        "200":
          description: Report list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [reports]
      summary: Create a report.
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreate"
      responses:
        "200":
          description: Created report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/reports/{reportId}:
    parameters:
      - $ref: "#/components/parameters/ReportId"
    get:
      tags: [reports]
      summary: Get report detail.
      operationId: getReport
      responses:
        "200":
          description: Report detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [reports]
      summary: Update report detail.
      operationId: updateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreate"
      responses:
        "200":
          description: Updated report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [reports]
      summary: Delete report.
      operationId: deleteReport
      responses:
        "204":
          description: Report deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/reports/{reportId}/status:
    parameters:
      - $ref: "#/components/parameters/ReportId"
    patch:
      tags: [reports]
      summary: Update report status.
      operationId: updateReportStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/reports/{reportId}/feedback:
    parameters:
      - $ref: "#/components/parameters/ReportId"
    patch:
      tags: [reports]
      summary: Update specialist feedback.
      operationId: updateReportFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackUpdate"
      responses:
        "200":
          description: Updated report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/reports/{reportId}/classification:
    parameters:
      - $ref: "#/components/parameters/ReportId"
    get:
      tags: [reports]
      summary: Get deterministic demo classification for a report.
      operationId: getReportClassification
      responses:
        "200":
          description: Classification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/reports/{reportId}/reclassify:
    parameters:
      - $ref: "#/components/parameters/ReportId"
    post:
      tags: [reports]
      summary: Recompute deterministic demo classification for a report.
      operationId: reclassifyReport
      responses:
        "200":
          description: Classification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ReportId:
      in: path
      name: reportId
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Invalid request payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        error:
          type: string
        detail:
          type: string
...
    ReportStatus:
      type: string
      enum: [DRAFT, ACTIVE, SUBMITTED, SENT, DONE, ERROR]
    ReportCreate:
      type: object
      required: [patient_id, doctor_id, target_organization_id]
      properties:
        patient_id:
          type: integer
        doctor_id:
          type: integer
        target_organization_id:
          type: integer
        status:
          $ref: "#/components/schemas/ReportStatus"
        authored_on:
          type: string
          format: date-time
          nullable: true
        note:
          type: string
        anamnesis:
          type: string
        mkn10_code:
          type: string
        severity:
          type: string
        feedback_specialist:
          type: string
        is_new_patient:
          type: boolean
        any_imaging_performed:
          type: boolean
        additional_imaging_planned:
          type: boolean
    ReportRead:
      allOf:
        - $ref: "#/components/schemas/ReportCreate"
        - type: object
          required: [id, status, status_cz, created_at, updated_at]
          properties:
            id:
              type: integer
            fhir_id:
              type: string
            status_cz:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/ReportStatus"
    FeedbackUpdate:
      type: object
      properties:
        feedback_specialist:
          type: string
        status:
          $ref: "#/components/schemas/ReportStatus"
    Classification:
      type: object
      additionalProperties: true
      properties:
        specialist_type:
          type: string
        specialist_reason:
          type: string
        severity:
          type: string
        severity_code:
          type: string
        severity_reason:
          type: string
...